Move to a separate class + use inbuilt methods for retrieval
